+++ lib/src/string_normalizer.dart
@@ -1,6 +1,7 @@
-import 'package:string_normalizer/src/utils.dart';
+import 'package:characters/characters.dart';
 
 import 'data/local.dart';
+import 'utils.dart';
 
 /// StringNormalizer extension
 extension StringNormalizerE on String {
@@ -17,10 +18,11 @@ extension StringNormalizerE on String {
 
 /// Convert all diacritical characters to ASCII characters.
 class StringNormalizer {
+  // Internal.
   StringNormalizer._();
 
   /// We do not need to initialize the data if it's not needed to use.
-  static Map<int, String>? _data;
+  static Map<String, String>? _data;
 
   /// Normalize the [text].
   ///
@@ -30,8 +32,17 @@ class StringNormalizer {
 
     StringBuffer result = StringBuffer();
 
-    for (final charCode in text.codeUnits) {
-      result.write(_data![charCode] ?? String.fromCharCode(charCode));
+    // Try to replace a whole character first then replace each code unit of a character.
+    for (final char in text.characters) {
+      if (_data!.containsKey(char)) {
+        result.write(_data![char]);
+      } else {
+        final codeUnits = char.codeUnits;
+        for (final codeUnit in codeUnits) {
+          final char = String.fromCharCode(codeUnit);
+          result.write(_data![char] ?? char);
+        }
+      }
     }
 
     return result.toString();